Simplifying the UI of the Application

This function simplifies the user interface of the app. It must be used inside a pmap function and be used with a corresponding vars variable built with tribble. The vars variable must have a column for id and choices, where id is the variable identifying the input and choices is the variable identifying the choices for the slider or selectize input. When used inside the pmap function, it automatically generates a selectize input or slider input depending on the variable type. For example, for date the input generated is a slider, whereas for type the variable generated is a select input.

There are no inputs required as shown in the example below, and the output returned is shown in the application.

selectize_input(id, label = id, choices = choices, multiple = multiple)

Arguments

id
This is the variable to identify the slider or selectize input, which should be stored in vars.

label
This variable is the label given to the slider or selectize input, the function will automatically generate the label from the specified id.

choices
This variable is the choices given to the slider or selectize input, which should be stored in vars.

multiple
This variable is about whether the input should have the option to select multiple variables, the function will automatically generate the response depending on the id.

Examples

library(purrr)
library(tibble)
library(coronaaus)
coronaaus::corona_aus

#> # A tibble: 6,230 x 7
#>    date       type      province                     lat  long cases cases_total
#>    <date>     <chr>     <chr>                      <dbl> <dbl> <int>       <dbl>
#>  1 2020-01-22 confirmed Australian Capital Territ… -35.5  149.     0           0
#>  2 2020-01-22 confirmed New South Wales            -33.9  151.     0           0
#>  3 2020-01-22 confirmed Northern Territory         -12.5  131.     0           0
#>  4 2020-01-22 confirmed Queensland                 -27.5  153.     0           0
#>  5 2020-01-22 confirmed South Australia            -34.9  139.     0           0
#>  6 2020-01-22 confirmed Tasmania                   -42.9  147.     0           0
#>  7 2020-01-22 confirmed Victoria                   -37.8  145.     0           0
#>  8 2020-01-22 confirmed Western Australia          -32.0  116.     0           0
#>  9 2020-01-22 death     Australian Capital Territ… -35.5  149.     0           0
#> 10 2020-01-22 death     New South Wales            -33.9  151.     0           0
#> # … with 6,220 more rows

vars <- tibble::tribble(~ id, ~ choices, 
"cities", unique(corona_aus$province), 
"type",  unique(corona_aus$type), 
"date", unique(corona_aus$date))
purrr::pmap(vars, coronaaus::selectize_input)

#> [[1]]
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="cities">Select a case cities</label>
#>   <div>
#>     <select id="cities" class="form-control" multiple="multiple"><option value="Australian Capital Territory">Australian Capital Territory</option>
#> <option value="New South Wales">New South Wales</option>
#> <option value="Northern Territory">Northern Territory</option>
#> <option value="Queensland">Queensland</option>
#> <option value="South Australia">South Australia</option>
#> <option value="Tasmania">Tasmania</option>
#> <option value="Victoria">Victoria</option>
#> <option value="Western Australia">Western Australia</option></select>
#>     <script type="application/json" data-for="cities">{}</script>
#>   </div>
#> </div>
#> 
#> [[2]]
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="type">Select a case type</label>
#>   <div>
#>     <select id="type" class="form-control"><option value="confirmed" selected>confirmed</option>
#> <option value="death">death</option>
#> <option value="recovered">recovered</option></select>
#>     <script type="application/json" data-for="type">{}</script>
#>   </div>
#> </div>
#> 
#> [[3]]
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="date">Select a case date</label>
#>   <input class="js-range-slider" id="date" data-type="double" data-min="1579651200000" data-max="1602115200000" data-from="1579651200000" data-to="1579651200000" data-step="86400000" data-grid="true" data-grid-num="10" data-grid-snap="false" data-prettify-separator="," data-prettify-enabled="true" data-keyboard="true" data-drag-interval="true" data-data-type="date" data-time-format="%F"/>
#> </div>
#>